apiextensions: unify multi- and mono-versioned test helpers
