+++ staging/src/k8s.io/apiextensions-apiserver/test/integration/versioning_test.go
@@ -19,14 +19,9 @@ package integration
 import (
 	"reflect"
 	"testing"
-	"time"
 
 	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/apiextensions-apiserver/test/integration/testserver"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/watch"
-	"k8s.io/client-go/dynamic"
 )
 
 func TestVersionedNamspacedScopedCRD(t *testing.T) {
@@ -43,7 +38,7 @@ func TestVersionedNamspacedScopedCRD(t *testing.T) {
 	}
 
 	ns := "not-the-default"
-	testSimpleVersionedCRUD(t, ns, noxuDefinition, dynamicClient)
+	testSimpleCRUD(t, ns, noxuDefinition, dynamicClient)
 }
 
 func TestVersionedClusterScopedCRD(t *testing.T) {
@@ -60,7 +55,7 @@ func TestVersionedClusterScopedCRD(t *testing.T) {
 	}
 
 	ns := ""
-	testSimpleVersionedCRUD(t, ns, noxuDefinition, dynamicClient)
+	testSimpleCRUD(t, ns, noxuDefinition, dynamicClient)
 }
 
 func TestStoragedVersionInNamespacedCRDStatus(t *testing.T) {
@@ -138,167 +133,5 @@ func testStoragedVersionInCRDStatus(t *testing.T, ns string, noxuDefinition *api
 	err = testserver.DeleteCustomResourceDefinition(crd, apiExtensionClient)
 	if err != nil {
 		t.Fatal(err)
-	}
-}
-
-func testSimpleVersionedCRUD(t *testing.T, ns string, noxuDefinition *apiextensionsv1beta1.CustomResourceDefinition, dynamicClient dynamic.Interface) {
-	noxuResourceClients := map[string]dynamic.ResourceInterface{}
-	noxuWatchs := map[string]watch.Interface{}
-	disbaledVersions := map[string]bool{}
-	for _, v := range noxuDefinition.Spec.Versions {
-		disbaledVersions[v.Name] = !v.Served
-	}
-	for _, v := range noxuDefinition.Spec.Versions {
-		noxuResourceClients[v.Name] = NewNamespacedCustomResourceVersionedClient(ns, dynamicClient, noxuDefinition, v.Name)
-
-		noxuWatch, err := noxuResourceClients[v.Name].Watch(metav1.ListOptions{})
-		if disbaledVersions[v.Name] {
-			if err == nil {
-				t.Errorf("expected the watch creation fail for disabled version %s", v.Name)
-			}
-		} else {
-			if err != nil {
-				t.Fatal(err)
-			}
-			noxuWatchs[v.Name] = noxuWatch
-		}
-	}
-	defer func() {
-		for _, w := range noxuWatchs {
-			w.Stop()
-		}
-	}()
-
-	for version, noxuResourceClient := range noxuResourceClients {
-		createdNoxuInstance, err := instantiateVersionedCustomResource(t, testserver.NewVersionedNoxuInstance(ns, "foo", version), noxuResourceClient, noxuDefinition, version)
-		if disbaledVersions[version] {
-			if err == nil {
-				t.Errorf("expected the CR creation fail for disabled version %s", version)
-			}
-			continue
-		}
-		if err != nil {
-			t.Fatalf("unable to create noxu Instance:%v", err)
-		}
-		if e, a := noxuDefinition.Spec.Group+"/"+version, createdNoxuInstance.GetAPIVersion(); e != a {
-			t.Errorf("expected %v, got %v", e, a)
-		}
-		for watchVersion, noxuWatch := range noxuWatchs {
-			select {
-			case watchEvent := <-noxuWatch.ResultChan():
-				if e, a := watch.Added, watchEvent.Type; e != a {
-					t.Errorf("expected %v, got %v", e, a)
-					break
-				}
-				createdObjectMeta, err := meta.Accessor(watchEvent.Object)
-				if err != nil {
-					t.Fatal(err)
-				}
-				// it should have a UUID
-				if len(createdObjectMeta.GetUID()) == 0 {
-					t.Errorf("missing uuid: %#v", watchEvent.Object)
-				}
-				if e, a := ns, createdObjectMeta.GetNamespace(); e != a {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-				createdTypeMeta, err := meta.TypeAccessor(watchEvent.Object)
-				if err != nil {
-					t.Fatal(err)
-				}
-				if e, a := noxuDefinition.Spec.Group+"/"+watchVersion, createdTypeMeta.GetAPIVersion(); e != a {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-				if e, a := noxuDefinition.Spec.Names.Kind, createdTypeMeta.GetKind(); e != a {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-			case <-time.After(5 * time.Second):
-				t.Errorf("missing watch event")
-			}
-		}
-
-		// Check get for all versions
-		for version2, noxuResourceClient2 := range noxuResourceClients {
-			// Get test
-			gottenNoxuInstance, err := noxuResourceClient2.Get("foo", metav1.GetOptions{})
-
-			if disbaledVersions[version2] {
-				if err == nil {
-					t.Errorf("expected the get operation fail for disabled version %s", version2)
-				}
-			} else {
-				if err != nil {
-					t.Fatal(err)
-				}
-
-				if e, a := version2, gottenNoxuInstance.GroupVersionKind().Version; !reflect.DeepEqual(e, a) {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-			}
-
-			// List test
-			listWithItem, err := noxuResourceClient2.List(metav1.ListOptions{})
-			if disbaledVersions[version2] {
-				if err == nil {
-					t.Errorf("expected the list operation fail for disabled version %s", version2)
-				}
-			} else {
-				if err != nil {
-					t.Fatal(err)
-				}
-				if e, a := 1, len(listWithItem.Items); e != a {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-				if e, a := version2, listWithItem.GroupVersionKind().Version; !reflect.DeepEqual(e, a) {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-				if e, a := version2, listWithItem.Items[0].GroupVersionKind().Version; !reflect.DeepEqual(e, a) {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-			}
-		}
-
-		// Delete test
-		if err := noxuResourceClient.Delete("foo", metav1.NewDeleteOptions(0)); err != nil {
-			t.Fatal(err)
-		}
-
-		listWithoutItem, err := noxuResourceClient.List(metav1.ListOptions{})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if e, a := 0, len(listWithoutItem.Items); e != a {
-			t.Errorf("expected %v, got %v", e, a)
-		}
-
-		for _, noxuWatch := range noxuWatchs {
-			select {
-			case watchEvent := <-noxuWatch.ResultChan():
-				if e, a := watch.Deleted, watchEvent.Type; e != a {
-					t.Errorf("expected %v, got %v", e, a)
-					break
-				}
-				deletedObjectMeta, err := meta.Accessor(watchEvent.Object)
-				if err != nil {
-					t.Fatal(err)
-				}
-				// it should have a UUID
-				createdObjectMeta, err := meta.Accessor(createdNoxuInstance)
-				if err != nil {
-					t.Fatal(err)
-				}
-				if e, a := createdObjectMeta.GetUID(), deletedObjectMeta.GetUID(); e != a {
-					t.Errorf("expected %v, got %v", e, a)
-				}
-
-			case <-time.After(5 * time.Second):
-				t.Errorf("missing watch event")
-			}
-		}
-
-		// Delete test
-		if err := noxuResourceClient.DeleteCollection(metav1.NewDeleteOptions(0), metav1.ListOptions{}); err != nil {
-			t.Fatal(err)
-		}
-
 	}
 }
